requirements.txt
apipkg==1.5
appdirs==1.4.4
asttokens==2.0.4
attrs==20.3.0
base58==2.1.0
bitarray==1.2.2
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
# version from branch used due to this error: https://github.com/eth-brownie/brownie/pull/1070
eth-brownie @ git+https://github.com/kilogold/brownie.git@034e00d8728ecda9e8db185d963cc6747e803bb9
eth-event==1.2.3
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.8.0
Flask==2.0.1
hexbytes==0.2.1
hypothesis==6.10.0
idna==2.10
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.7.0a1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
packaging==20.9
parsimonious==0.8.1
pathspec==0.8.1
pluggy==0.13.1
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.8.0
py==1.10.0
py-solc-ast==1.2.8
py-solc-x==1.1.0
pycryptodome==3.10.1
Pygments==2.8.1
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pythx==1.6.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rlp==1.2.0
semantic-version==2.8.5
six==1.15.0
sortedcontainers==2.3.0
toml==0.10.2
toolz==0.11.1
tqdm==4.60.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
varint==1.0.2
vvm==0.1.0
vyper==0.2.12
wcwidth==0.2.5
web3==5.18.0
websockets==8.1
Werkzeug==2.0.1
